Add GitHub Action to build images on PRs
Building images on PRs will help to ease the testing of such images
The images pushed use the commit hash as the tag.
Only PRs coming from the same repo will trigger the build+push of
images.
It's possible to manually force a build+push of images in a branch
(without a PR) through a dispatch workflow using a branch in the
repo.